requirements.txt
boto==2.49.0
boto3==1.14.28
botocore==1.17.28
docutils==0.15.2
jmespath==0.10.0
python-dateutil==2.8.1
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.10
boto==2.49.0
boto3==1.14.28
botocore==1.17.28
docutils==0.15.2
jmespath==0.10.0
python-dateutil==2.8.1
PyYAML==5.3.1
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.10
